3 Different Models on the same dataset
This code is the code that I use to conduct research regarding the accuracy of a model to determine whether the text is sarcasm or not.
I hope on future works that this code can help research related to LSTM, BILSTM and BERT models in the future.
Apart from that, this code is also expected to be useful for carrying out research on word analysis, whether it is classifying ordinary words or sarcasm.